upload

+++ app/Http/Controllers/ProjectsController.php
@@ -6,15 +6,28 @@ use Illuminate\Http\Request;
 
 use App\Project;
 
+use Mail;
+
+use App\Mail\ProjectCreated;
+
 class ProjectsController extends Controller
 {
+    
+    public function __construct()
+    
+    {
+        $this->middleware('auth');
+    }
+
     public function index()
     {
         // $project= App\Project::all();
 
-        $projects= Project::all();
+        // $projects= Project::all();
             // return $projects;
 
+        // Get all the projects for the authenticated user
+        $projects = Project::where('owner_id',auth()->id())->get();
         return view('projects.index',compact('projects'));
 
     }
@@ -22,6 +35,18 @@ class ProjectsController extends Controller
     public function show($id)
     {
         $projectdata = Project::findOrFail($id);
+        
+        // if($projectdata->owner_id !== auth()->id()) {
+        //     abort(403);
+        // }
+
+        // abort_if()
+
+        $this->authorize('update',$projectdata);
+
+        // abort_unless()
+        // abort_unless(\Gate::allows('update',$projectdata),403);
+
         return view('projects.show',compact('projectdata'));
     }
 
@@ -41,10 +66,10 @@ class ProjectsController extends Controller
             'description' => ['required', 'min:3']
             ]);
 
-        Project::create([
+        $project= Project::create([
             'title'=>request('title'),
             'description'=>request('description')
-        ]);
+        ]+ ['owner_id' => auth()->id()]);
 
         // $project=new Project();
         
@@ -52,8 +77,16 @@ class ProjectsController extends Controller
         // $project->description= request('description');
 
         // $project->save();
-
-return redirect('/projects');
+        
+        // Mailables
+        // Mail::to('m.azizulcse@gmail.com')->send(
+        //     new ProjectCreated($project)
+        //     );
+        // $project->owner
+        Mail::to($project->owner->email)->send(
+            new ProjectCreated($project)
+            );
+        return redirect('/projects');
 
     }
     
@@ -81,11 +114,13 @@ return redirect('/projects');
      // dd(request()->all());
      
      $project = Project::findOrFail($id);
-     
-     $project->title= request('title');
-     $project->description= request('description');
 
-     $project->save();
+     $this->authorize('update',$project);
+     
+     // $project->title= request('title');
+     // $project->description= request('description');
+     // $project->save();
+     $project->update($this->validateProject());
 
      return redirect('/projects');
 
@@ -95,9 +130,20 @@ return redirect('/projects');
     {
         // dd('hello');
         // Project::findOrFail($id)->delete();
+        
          $project = Project::findOrFail($id);
+         $this->authorize('update',$project);
+
          $project->delete();
 
         return redirect('/projects');
+    }
+
+    protected function validateProject()
+    {
+        return request()->validate([
+            'title' => ['required','min:3'],
+            'description' => ['required','min:3']
+            ]);
     }
 }
